breakout I trade document
