+++ db/trade/index.ts
@@ -1,14 +1,5 @@
-import mongoose, { Schema } from 'mongoose';
-import { BuyOrSell, StockOrOption } from '../types';
+import { Schema } from 'mongoose';
 
-export interface ITradeDocument extends mongoose.Document {
-  price: number;
-  quantity: number;
-  side: BuyOrSell;
-  symbol: string;
-  timestamp: Date;
-  type: StockOrOption;
-}
-
+export * from './ITradeDocument';
 export * from './OptionTrade';
 export * from './StockTrade';
